ui: Fix bug preventing service providers from saving
When making an API call to create/edit a service, we needed a way to
know if the user performing the action has used 'my-providers' or
'providers' field in order to handle business logic accordingly.
Since serialize() method does not have access to the user's role,
we assign the role to the dummy model property 'user_role' after the
model has been loaded.