ui: Simplify my-service/my-service-versions functionality
In order to reduce code complexity, we used a different approach:
When we want to fetch the services that a serviceadmin owns, instead
of using /my-services endpoint, we use /services endpoint and filter
serveri-side to get the services that belong to ServiceAdminships where
the admin is the serviceadmin. We then filter client-side to get only
the approved adminships.