ui: Fix bug with deleted Resource Adminship remaining in list
In list view, a deleted Resource Adminship instance would remain visible
in list, until the user would refresh the page.
This was due to getModel() in list view, that would interfere with
rerendering the list after delete.
I removed getModel() so the user pending Resource Adminships are shown
in list.
The actions concerning these pending applications are hidden.